+++ src/features/godbolt/lib/func/func-wasm/func-compile.ts
