+++ src/features/godbolt/lib/func/func-wasm/funcfiftlib.js
