+++ src/features/godbolt/lib/func/stdlib.ts
