+++ src/features/tasm/hooks/usePlaygroundTrace.ts
@@ -1,7 +1,7 @@
 import {useEffect, useState} from "react"
 import {createTraceInfoPerTransaction, type TraceInfo} from "ton-assembly/dist/trace"
 
-import type {AssemblyExecutionResult} from "@features/txTrace/lib/asm/assemblyExecutor.ts"
+import type {AssemblyExecutionResult} from "@features/tasm/lib/executor.ts"
 import type {ExitCode} from "@features/txTrace/lib/traceTx"
 import {findExitCode} from "@features/txTrace/lib/traceTx"
