+++ src/features/tasm/lib/index.ts
@@ -1,0 +1,2 @@
+export * from "./types"
+export * from "./documentation.ts"
