--- src/features/txTrace/lib/asm/index.ts
@@ -1,2 +1,0 @@
-export * from "./types"
-export * from "./asm-documentation"
